Add Product Module

+++ model/product.js
@@ -9,7 +9,7 @@ module.exports = (sequelize,DataTypes)=>{
             type:DataTypes.TEXT
         },
         weight:{
-            type:DataTypes.INTEGER
+            type:DataTypes.DECIMAL(10,2)
         },
         category_id:{
             type:DataTypes.INTEGER
